requirements.txt
Flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0 